// src/interfaces/Actions/LoopThroughTable.tsx
// Generated by https://quicktype.io

export interface LoopThroughTable {
    Type:               string;
    ViewNameFriendly:   string;
    WhereClause:        WhereClause[];
    DistinctColumn:     null;
    Sorting:            Sorting[];
    LoopActionSet:      LoopActionSet;
    Notes:              string;
    ActionId:           number;
    Disabled:           boolean;
    InValid:            boolean;
    ValidationMessages: any[];
    Metadata:           null;
}

export interface LoopActionSet {
    Name:         string;
    ActionSetId:  string;
    Key:          string;
    Scope:        number;
    Actions:      any[];
    Notes:        string;
    IsInAppStore: boolean;
}

export interface Sorting {
    Field:              string;
    IsAscending:        boolean;
    ValidationMessage:  null;
    ValidationMessages: any[];
}

export interface WhereClause {
    Operation:         number | null;
    DateField:         null;
    Source:            Source | null;
    AdditionalSources: null;
    DateWhere:         null;
}

export interface Source {
    ValueType:         number;
    Value:             number | string;
    Property:          string;
    ChildProperty:     null;
    ValidationMessage: null;
    FontStyles:        null;
}